requirements.txt
Flask
Werkzeug
requests
openai
python-dotenv
